Interactive.Plot

Description

Simple interactive rendering of plots

Run a Plot

runPlot

Arguments

:: PlotOpts

-> Maybe String
title

-> Maybe Image
description box

-> [Series]
series data

-> IO ()

Display fixed plot and title interactively.

See runPlotDynamic for more control.

runPlotAuto

Arguments

:: PlotOpts

-> Maybe String
title

-> Maybe Image
description box

-> [AutoSeries]
uninitialized series data

-> IO ()

Display fixed plot and title interactively, filling in default values.

See runPlotDynamic for more control.

Animated

animatePlot

Arguments

:: PlotOpts

-> Double
update rate (frames per second)

-> Maybe String
title

-> Maybe Image
description box

-> [[Series]]
list of series data (potentially infinite)

-> IO ()

Display a series of plots ([SeriesF]) with a time delay between each one. Will quit when the last plot is displayed. Use lastForever on the input list to repeat the last item indefinitely, or cycle to cycle through the list forever.

Note that this behavior is pretty simple; more advanced functionality can be achieved with runPlotDynamic directly.

lastForever :: [a] -> [a]

Handy function to use with animatePlot to extend the last frame into eternity.

animatePlotFunc

Arguments

:: PlotOpts

-> Maybe String
title

-> Maybe Image
description box

-> (Double -> Maybe [Series])
function from time to plot. will quit as soon as Nothing is returned.

-> IO ()

Animate (according to the framerate in the PlotOpts) a function Double -> Maybe [Series], where the input is the current time in seconds and the output is the plot to display at that time. Will quit as soon as Nothing is given.

Remember to give a PlotOpts with a Just framerate.

This is a simple wrapper over animatePlotMoore with a stateless function. For more advanced functionality, use animatePlotMoore or runPlotDynamic directly.

animatePlotMoore

Arguments

:: PlotOpts

-> Maybe String
title

-> Maybe Image
description box

-> Moore [Series]
moore machine representing progression of plot from an initial state

-> IO ()

Animate (according to the framerate in the PlotOpts) a "Moore machine" description of a plot evolving over time with some initial state.

Remember to give a PlotOpts with a Just framerate.

For a simplified version of a stateless function, see animatePlotFunc. This is implemented in terms of runPlotDynamic, but the representation of an animation in terms of a moore machine is powerful enough to represent a very general class of animations.

data Moore a

Used for animatePlotMoore to specify how a plot evolves over time with some initial state.

Constructors

Moore

Fields

moInitVal :: Maybe a

moInitState :: s

initial state of plot

moUpdate :: Double -> s -> IO (Maybe (a, s))

Given change in time since last render and old state, return new plot and state. Return Nothing to quit.

Options

data PlotOpts

Options used for running the plot interactively in a terminal.

Constructors

PO

Fields

_poTermRatio :: Double

character width ratio of terminal (H/W). Default is 2.1.

_poAspectRatio :: Maybe Double

plot aspect ratio (H/W). Use Nothing for automatic. Default is Just 1.

_poXRange :: Maybe (Range Double)

X Range. Use Nothing for automatic. Default is Nothing.

_poYRange :: Maybe (Range Double)

Y Range. Use Nothing for automatic. Default is Nothing.

_poAutoMethod :: Maybe StdGen

How to fill in missing values when run using runPlotAuto. Nothing for IO, Just for deterministic seed. Ignored when using runPlot. Default is an arbitrarily selected seed.

_poHelp :: Bool

Whether or not to show help box initially. Box can always be toggled with ?. (Default is True)

_poFramerate :: Maybe Double

Updates per second; Nothing for no updates. Use poDelay to treat this as a microsecond delay instead. (default: Nothing)

poTermRatio :: Lens' PlotOpts Double

poAspectRatio :: Lens' PlotOpts (Maybe Double)

poXRange :: Lens' PlotOpts (Maybe (Range Double))

poYRange :: Lens' PlotOpts (Maybe (Range Double))

poRange :: Lens' PlotOpts (Maybe (Range Double), Maybe (Range Double))

Lens into a PlotOpts getting its range X and range Y settings.

poAutoMethod :: Lens' PlotOpts (Maybe StdGen)

poHelp :: Lens' PlotOpts Bool

poFramerate :: Lens' PlotOpts (Maybe Double)

poDelay :: Lens' PlotOpts (Maybe Int)

Lens into microsecond delay between frames, specified by a PlotOpts.

defaultPlotOpts :: PlotOpts

Sensible defaults for most terminals.

Construct Series

data Auto a

Used to specify fields in PointStyle and SeriesF: Use Auto for automatic inference, and Given to provide a specific value.

Its Semigroup instance keeps the last Given.

Constructors

Auto

Given a

type Series = SeriesF Identity

Data for a single series: contains the coordinate map with the point style for the series.

type AutoSeries = SeriesF Auto

A version of SeriesF where you can leave the marker or color blank, to be automatically inferred.

data SeriesF f

A parameterized version of SeriesF to unify functions in Interactive.Plot.Series.

Mainly you will be using either SeriesF or AutoSeries.

Constructors

Series

Fields

_sItems :: Map Double (Set Double)

A map of x positions to y points at that position

_sStyle :: PointStyleF f

The style of points. For SeriesF, this is PointStyle; for AutoSeries, this is AutoPointStyle.

sItems :: Lens' (SeriesF f) (Map Double (Set Double))

Getter/setter lens to the items field of a SeriesF

sStyle :: Lens' (SeriesF f) (PointStyleF f)

Getter/setter lens to the style field of a SeriesF

sStyle :: Lens SeriesF PointStyle
sStyle :: Lens AutoSeries AutoPointStyle

Making common serieses

listSeries :: Foldable t => t Double -> PointStyleF f -> SeriesF f

Construct a series from any foldable container of y-values. The x-values are automatically assigned to 0, 1, 2, 3 ... etc.

Note that this is polymorphic over both PointStyle and AutoPointStyle:

listSeries :: Foldable t => t Double -> PointStyle -> SeriesF
listSeries :: Foldable t => t Double -> AutoPointStyle -> AutoSeries

tupleSeries :: Foldable t => t (Double, Double) -> PointStyleF f -> SeriesF f

Construct a series from any foldable container of x-y tuples.

Note that this is polymorphic over both PointStyle and AutoPointStyle:

tupleSeries :: Foldable t => t (Double, Double) -> PointStyle -> SeriesF
tupleSeries :: Foldable t => t (Double, Double) -> AutoPointStyle -> AutoSeries

funcSeries :: Foldable t => (Double -> Double) -> t Double -> PointStyleF f -> SeriesF f

Construct a series from a function x to y, given a foldable container of x values.

Note that this is polymorphic over both PointStyle and AutoPointStyle:

funcSeries :: Foldable t => (Double -> Double) -> t Double -> PointStyle -> SeriesF
funcSeries :: Foldable t => (Double -> Double) -> t Double -> AutoPointStyle -> AutoSeries

enumRange

Arguments

:: Fractional a

=> Int
Number of points

-> Range a
Range to generate the points over

-> [a]

enumRange n (R a b) generates a list of n equally spaced values between a and b.

toCoordMap :: Eq a => Set (Coord a) -> Map a (Set a)

Turn a set of coordinates into a map of x's to the y's found in the set.

Note that this forms an isomorphism with fromCoordMap.

fromCoordMap :: Map a (Set a) -> Set (Coord a)

Convert a map of x's to y's into a set of x-y coordinates.

Note that this forms an isomorphism with toCoordMap.

Series from AutoSeroes

fromAutoSeries :: [AutoSeries] -> [Series]

Turn an AutoSeries into a SeriesF, assigning styles from a pre-specified "shuffled" order.

fromAutoSeriesIO :: [AutoSeries] -> IO [Series]

Turn an AutoSeries into a SeriesF, drawing styles randomly in IO.

fromAutoSeries_ :: StdGen -> [AutoSeries] -> [Series]

Turn an AutoSeries into a SeriesF, shuffling the default styles in a deterministic way from a given seed.

Types

type PointStyle = PointStyleF Identity

Specification of a style for a point.

Construct this wiht the PointStyle pattern synonym.

pattern PointStyle :: Char -> Color -> PointStyle

Pattern synonym/constructor for PointStyle.

This comes with two record fields, _psMarker and _psColor.

_psMarker :: PointStyle -> Char

_psColor :: PointStyle -> Color

type AutoPointStyle = PointStyleF Auto

A version of PointStyle where you can leave the marker or color blank, to be automatically inferred.

You can construct this with the PointStyleF constructor.

It has a very convenient 'Monoid'/'Semigroup' instance: mempty gives a PointStyle where every field is Auto, and <> combines PointStyles field-by-field, keeping the last Given.

data PointStyleF f

A parameterized version of PointStyle to unify functions in Interactive.Plot.Series.

Mainly you will be using either PointStyle or AutoPointStyle.

Constructors

PointStyleF

Fields

_psMarkerF :: f Char

Marker character.

_psColorF :: f Color

Marker color.

psMarker :: Lens' PointStyle Char

psColor :: Lens' PointStyle Color

data Coord a

An ordered pair in a.

Constructors

C

Fields

_cX :: a

Access x.

_cY :: a

Access y.

cX :: Lens' (Coord a) a

Getter/setter lens to the x position in a Coord.

cY :: Lens' (Coord a) a

Getter/setter lens to the x position in a Coord.

data Range a

A specification for a range. Using R, contains the minimum and maximum. Using RAbout, contains the midpoint and size.

Constructors

R

Fields

_rMin :: a

Minimum of range.

_rMax :: a

Maximum of range.

Bundled Patterns

pattern RAbout :: Fractional a => a -> a -> Range a
An alternative "constructor" for R, which takes a midpoint and size instead of a min and max.

This comes with record fields, _rMid and _rSize'.

_rMid :: Range a -> Fractional a -> a

_rSize' :: Range a -> Fractional a -> a

rMin :: Lens' (Range a) a

Getter/setter lens to the minimum value in a Range.

rMax :: Lens' (Range a) a

Getter/setter lens to the maximum value in a Range.

rSize :: Fractional a => Lens' (Range a) a

Lens into the size of a Range Modifying this size results in a scaling about the midpoint of the range.

view rSize (R 2 4)
-- 2
over rSize (* 2) (R 2 4)
-- R 1 5

rMid :: Fractional a => Lens' (Range a) a

Lens into the midpoint of a Range. Modifying this midpoint shifts the range to a new midpoint, preserving the size.

view rMid (R 2 4)
-- 3
over rMid (+ 3) (R 2 4)
-- R 5 7

_rSize :: Num a => Range a -> a

Gets the size of a Range.

A version of _rSize' that works for any instance of Num.