Interactive.Plot.Core

Description

Core rendering functionality for the library.

Documentation

data Coord a

An ordered pair in a.

Constructors

C

Fields

_cX :: a

Access x. For getter/setter lenses, see cX.

_cY :: a

Access y. For getter/setter lenses, see cY.

cX :: forall a. Lens' (Coord a) a

cY :: forall a. Lens' (Coord a) a

data Range a

A specification for a range. Using R, contains the minimum and maximum. Using RAbout, contains the midpoint and size.

Constructors

R

Fields

_rMin :: a

Minimum of range. For getter/setter lens, see rMin.

_rMax :: a

Maximum of range. For getter/setter lens, see rMax.

Bundled Patterns

pattern RAbout :: Fractional a => a -> a -> Range a
An alternative "constructor" for R, which takes a midpoint and size instead of a min and max.

rMin :: forall a. Lens' (Range a) a

rMax :: forall a. Lens' (Range a) a

rSize :: Fractional a => Lens' (Range a) a

Lens into the size of a Range Modifying this size results in a scaling about the midpoint of the range.

view rSize (R 2 4)
-- 2
over rSize (* 2) (R 2 4)
-- R 1 5

rMid :: Fractional a => Lens' (Range a) a

Lens into the midpoint of a Range. Modifying this midpoint shifts the range to a new midpoint, preserving the size.

view rMid (R 2 4)
-- 3
over rMid (+ 3) (R 2 4)
-- R 5 7

_rSize :: Num a => Range a -> a

A version of _rSize that works for any instance of Num.

data Auto a

Constructors

Auto

Given a

type PointStyle = PointStyleF Identity

Specification of a style for a point.

Construct this wiht the PointStyle pattern synonym.

pattern PointStyle :: Char -> Color -> PointStyle

Pattern synonym/constructor for PointStyle.

data PointStyleF f

A parameterized version of PointStyle to unify functions in Interactive.Plot.Series.

Mainly you will be using either PointStyle or AutoPointStyle.

Constructors

PointStyleF

Fields

_psMarkerF :: f Char

Marker cahracter. For getter/setter lens, see psMarker.

_psColorF :: f Color

Marker color. For getter/setter lens, see psColor.

type AutoPointStyle = PointStyleF Auto

A version of PointStyle where you can leave the marker or color blank, to be automatically inferred.

You can construct this with the PointStyleF constructor.

psMarker :: Lens' PointStyle Char

psColor :: Lens' PointStyle Color

_psMarker :: PointStyle -> Char

_psColor :: PointStyle -> Color

type Series = SeriesF Identity

Data for a single series: contains the coordinate map with the point style for the series.

data SeriesF f

A parameterized version of SeriesF

Constructors

Series

Fields

_sItems :: Map Double (Set Double)

_sStyle :: PointStyleF f

type AutoSeries = SeriesF Auto

A version of SeriesF where you can leave the marker or color blank, to be automatically inferred.

sItems :: forall f. Lens' (SeriesF f) (Map Double (Set Double))

sStyle :: forall f f. Lens (SeriesF f) (SeriesF f) (PointStyleF f) (PointStyleF f)

toCoordMap :: Eq a => Set (Coord a) -> Map a (Set a)

Turn a set of coordinates into a map of x's to the y's found in the set.

Note that this forms an isomorphism with fromCoordMap.

fromCoordMap :: Map a (Set a) -> Set (Coord a)

Convert a map of x's to y's into a set of x-y coordinates.

Note that this forms an isomorphism with toCoordMap.

data Alignment

Alignment specification.

Constructors

ALeft

ACenter

ARight

data PlotOpts

Options used for running the plot interactively in a terminal.

Constructors

PO

Fields

_poTermRatio :: Double

character width ratio of terminal (H/W). Default is 2.1.

_poAspectRatio :: Maybe Double

plot aspect ratio (H/W). Use Nothing for automatic. Default is Just 1.

_poXRange :: Maybe (Range Double)

X Range. Use Nothing for automatic. Default is Nothing.

_poYRange :: Maybe (Range Double)

Y Range. Use Nothing for automatic. Default is Nothing.

_poAutoMethod :: Maybe StdGen

How to fill in missing values when run using runPlotAuto. Nothing for IO, Just for deterministic seed. Ignored when using runPlot. Default is an arbitrarily selected seed.

poTermRatio :: Lens' PlotOpts Double

poAspectRatio :: Lens' PlotOpts (Maybe Double)

poXRange :: Lens' PlotOpts (Maybe (Range Double))

poYRange :: Lens' PlotOpts (Maybe (Range Double))

poRange :: Lens' PlotOpts (Maybe (Range Double), Maybe (Range Double))

Lens into a PlotOpts getting its range X and range Y settings.

poAutoMethod :: Lens' PlotOpts (Maybe StdGen)

defaultPlotOpts :: PlotOpts

Sensible defaults for most terminals.

renderPlot

Arguments

:: Coord (Range Int)
display region

-> Coord (Range Double)
plot axis range

-> [Series]

-> [Image]

Render serieses based on a display region and plot axis ranges.

Internal

plotRange

Arguments

:: PlotOpts

-> Coord (Range Int)
display region

-> [Series]
Points

-> Coord (Range Double)
actual plot axis range

Compute plot axis ranges based on a list of points and the size of the display region.

newtype OrdColor

Newtype wrapper providing an Ord instance for Color.

Constructors

OC

Fields

getOC :: Color

renderPoint :: PointStyle -> Image

Render a single according to a PointStyle.