Interactive.Plot.Series

Description

Create common serieses.

Documentation

data AutoPointStyle

A version of PointStyle where you can leave the marker or color blank, to be automatically inferred.

Constructors

APS

Fields

_apsMarker :: Maybe Char

_apsColor :: Maybe Color

apsMarker :: Lens' AutoPointStyle (Maybe Char)

apsColor :: Lens' AutoPointStyle (Maybe Color)

data AutoSeries

A version of Series where you can leave the marker or color blank, to be automatically inferred.

Constructors

AS

Fields

_asItems :: Map Double (Set Double)

_asStyle :: AutoPointStyle

asItems :: Lens' AutoSeries (Map Double (Set Double))

asStyle :: Lens' AutoSeries AutoPointStyle

defaultStyles :: Set PointStyle

A set of default point styles

Create a Series from an AutoSeries.

fromAutoSeries :: [AutoSeries] -> [Series]

Turn an AutoSeries into a Series, assigning styles from a pre-specified "shuffled" order.

fromAutoSeriesIO :: [AutoSeries] -> IO [Series]

Turn an AutoSeries into a Series, drawing styles randomly in IO.

fromAutoSeries_ :: StdGen -> [AutoSeries] -> [Series]

Turn an AutoSeries into a Series, shuffling the default styles in a deterministic way from a given seed.

Create common Series

listSeries :: Foldable t => t Double -> AutoPointStyle -> AutoSeries

Construct a series from any foldable container of y-values.

tupleSeries :: Foldable t => t (Double, Double) -> AutoPointStyle -> AutoSeries

Construct a series from any foldable container of x-y tuples.

autoSeries :: Series -> AutoSeries

Convert from a Series back into an AutoSeries with settings given.

funcSeries :: Foldable t => (Double -> Double) -> t Double -> AutoPointStyle -> AutoSeries

Construct a series from a function x to y, given a foldable container of x values.

enumRange

Arguments

:: Fractional a

=> Int
Number of points

-> Range a
Range to generate the points over

-> [a]

enumRange n (R a b) generates a list of n equally spaced values between a and b.