Interactive.Plot.Series

Description

Create common serieses.

Documentation

type Series = SeriesF Identity

Data for a single series: contains the coordinate map with the point style for the series.

type AutoSeries = SeriesF Auto

A version of SeriesF where you can leave the marker or color blank, to be automatically inferred.

data SeriesF f

A parameterized version of SeriesF to unify functions in Interactive.Plot.Series.

Mainly you will be using either SeriesF or AutoSeries.

Constructors

Series

Fields

_sItems :: Map Double (Set Double)

A map of x positions to y points at that position

_sStyle :: PointStyleF f

The style of points. For SeriesF, this is PointStyle; for AutoSeries, this is AutoPointStyle.

type PointStyle = PointStyleF Identity

Specification of a style for a point.

Construct this wiht the PointStyle pattern synonym.

type AutoPointStyle = PointStyleF Auto

A version of PointStyle where you can leave the marker or color blank, to be automatically inferred.

You can construct this with the PointStyleF constructor.

It has a very convenient Monoid instance: mempty gives a PointStyle where every field is Auto, and <> combines PointStyles field-by-field, keeping the last Given.

data PointStyleF f

A parameterized version of PointStyle to unify functions in Interactive.Plot.Series.

Mainly you will be using either PointStyle or AutoPointStyle.

Constructors

PointStyleF

Fields

_psMarkerF :: f Char

Marker character.

_psColorF :: f Color

Marker color.

Create common SeriesF

listSeries :: Foldable t => t Double -> PointStyleF f -> SeriesF f

Construct a series from any foldable container of y-values. The x-values are automatically assigned to 0, 1, 2, 3 ... etc.

Note that this is polymorphic over both PointStyle and AutoPointStyle:

listSeries :: Foldable t => t Double -> PointStyle -> SeriesF
listSeries :: Foldable t => t Double -> AutoPointStyle -> AutoSeries

tupleSeries :: Foldable t => t (Double, Double) -> PointStyleF f -> SeriesF f

Construct a series from any foldable container of x-y tuples.

Note that this is polymorphic over both PointStyle and AutoPointStyle:

tupleSeries :: Foldable t => t (Double, Double) -> PointStyle -> SeriesF
tupleSeries :: Foldable t => t (Double, Double) -> AutoPointStyle -> AutoSeries

funcSeries :: Foldable t => (Double -> Double) -> t Double -> PointStyleF f -> SeriesF f

Construct a series from a function x to y, given a foldable container of x values.

Note that this is polymorphic over both PointStyle and AutoPointStyle:

funcSeries :: Foldable t => (Double -> Double) -> t Double -> PointStyle -> SeriesF
funcSeries :: Foldable t => (Double -> Double) -> t Double -> AutoPointStyle -> AutoSeries

enumRange

Arguments

:: Fractional a

=> Int
Number of points

-> Range a
Range to generate the points over

-> [a]

enumRange n (R a b) generates a list of n equally spaced values between a and b.

Create a SeriesF from an AutoSeries.

fromAutoSeries :: [AutoSeries] -> [Series]

Turn an AutoSeries into a SeriesF, assigning styles from a pre-specified "shuffled" order.

fromAutoSeriesIO :: [AutoSeries] -> IO [Series]

Turn an AutoSeries into a SeriesF, drawing styles randomly in IO.

fromAutoSeries_ :: StdGen -> [AutoSeries] -> [Series]

Turn an AutoSeries into a SeriesF, shuffling the default styles in a deterministic way from a given seed.

defaultStyles :: Set PointStyle

A set of default point styles