Interactive.Plot.Series

Description

Create common serieses.

Documentation

defaultStyles :: Set PointStyle

A set of default point styles

Create a SeriesF from an AutoSeries.

fromAutoSeries :: [AutoSeries] -> [Series]

Turn an AutoSeries into a SeriesF, assigning styles from a pre-specified "shuffled" order.

fromAutoSeriesIO :: [AutoSeries] -> IO [Series]

Turn an AutoSeries into a SeriesF, drawing styles randomly in IO.

fromAutoSeries_ :: StdGen -> [AutoSeries] -> [Series]

Turn an AutoSeries into a SeriesF, shuffling the default styles in a deterministic way from a given seed.

Create common SeriesF

listSeries :: Foldable t => t Double -> AutoPointStyle -> AutoSeries

Construct a series from any foldable container of y-values.

tupleSeries :: Foldable t => t (Double, Double) -> AutoPointStyle -> AutoSeries

Construct a series from any foldable container of x-y tuples.

autoSeries :: Series -> AutoSeries

Convert from a SeriesF back into an AutoSeries with settings given.

funcSeries :: Foldable t => (Double -> Double) -> t Double -> AutoPointStyle -> AutoSeries

Construct a series from a function x to y, given a foldable container of x values.

enumRange

Arguments

:: Fractional a

=> Int
Number of points

-> Range a
Range to generate the points over

-> [a]

enumRange n (R a b) generates a list of n equally spaced values between a and b.